updated video
hello friend
I saw your webpage. It is nice but the video which was there was not that much nice. So I suggested a video and edited the code for that video. So if you like that then please merge my code. I will be very happy you contribute to your project.